+++ client/src/components/index.js
@@ -17,7 +17,7 @@ export { Blob } from 'components/Cards/InfoCard/Blob'
 export { InfoCard } from 'components/Cards/InfoCard/InfoCard'
 export { ArrowBtnR } from 'components/Cards/MiniCard/ArrowBtnR'
 export { MiniCard } from 'components/Cards/MiniCard/MiniCard'
-export { Chart } from 'components/Charts/Chart'
+export { ChartBar } from 'components/Charts/ChartBar'
 export { default as FormBlog } from 'components/Forms/FormBlog'
 export { default as FormOffer } from 'components/Forms/FormOffer'
 export { LateralBar } from 'components/LateralBar/LateralBar'
