+++ client/src/main.jsx
@@ -6,7 +6,7 @@ import { router } from './router'
 import { Provider } from 'react-redux'
 import { store } from './redux/store/store'
 import { ClerkProvider } from '@clerk/clerk-react'
-import { AppContextProvider } from './pages/Account/OnboardingContext'
+import { AppContextProvider } from "./pages/Account/OnBoarding/AppContextProvider"
 
 // @ts-ignore
 const CLERK_PUB_KEY = import.meta.env.VITE_CLERK_PUB_KEY
