+++ client/src/pages/Account/OnBoarding/AppContext.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+// @ts-ignore
+
+export const AppContext = React.createContext();
