+++ client/src/pages/Account/steps/Step3.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react'
-import { AppContext } from '../OnboardingContext'
+import { AppContext } from "../OnBoarding/AppContext"
 import { ButtonContainerLg } from '../ButtonContainerLg'
 import { preferencesData } from '@/constants/preferencesData'
