+++ client/src/pages/Form/ProfileImageForm.jsx
@@ -1,0 +1,32 @@
+import { useState } from 'react'
+import { Image } from 'cloudinary-react'
+import { useUpload } from '@/hooks/upload'
+
+export function ProfileImageForm() {
+  const [file, setFile] = useState(null)
+  const { loading, data } = useUpload(file)
+  const image = data?.public_id || null
+
+  const handleChange = e => {
+    const file = e.target.files[0]
+    setFile(file)
+  }
+
+  return (
+    <div>
+      {!loading && image ? (
+        <div className="aspect-s w-24 overflow-hidden rounded-lg">
+          <Image cloudName="dvzhqzjkm" publicId={image} />
+        </div>
+      ) : (
+        <div className="aspect-square w-24 rounded-lg bg-slate-200" />
+      )}
+      <input
+        accept="image/*"
+        type="file"
+        name="image"
+        onChange={handleChange}
+      />
+    </div>
+  )
+}
