+++ client/src/pages/Form/UserForm.jsx
@@ -1,0 +1,4 @@
+// @ts-nocheck
+
+export const inputCls =
+  'rounded-md border border-slate-200 p-3 outline-none focus:ring-2'
