+++ client/src/pages/index.js
@@ -1,6 +1,7 @@
 export { ButtonContainerLg } from 'pages/Account/ButtonContainerLg'
-export { Onboarding } from 'pages/Account/Onboarding'
-export { AppContext, AppContextProvider } from 'pages/Account/OnboardingContext'
+export { AppContext } from 'pages/Account/OnBoarding/AppContext'
+export { AppContextProvider } from 'pages/Account/OnBoarding/AppContextProvider'
+export { Onboarding } from 'pages/Account/OnBoarding/Onboarding'
 export { Payment } from 'pages/Account/Payment'
 export { Step1 } from 'pages/Account/steps/Step1'
 export { Step2 } from 'pages/Account/steps/Step2'
@@ -9,9 +10,9 @@ export { Step4 } from 'pages/Account/steps/Step4'
 export { Step6 } from 'pages/Account/steps/Step6'
 export { AuthLayout } from 'pages/Auth/AuthLayout'
 export { ForgotPass } from 'pages/Auth/ForgotPass'
-export { LoginBack } from 'pages/Auth/Login.back'
+export { Login.back } from 'pages/Auth/Login.back'
 export { Login } from 'pages/Auth/Login'
-export { RegisterBack } from 'pages/Auth/Register.back'
+export { Register.back } from 'pages/Auth/Register.back'
 export { Register } from 'pages/Auth/Register'
 export { Companies } from 'pages/Community/Companies/Companies'
 export { default as Media } from 'pages/Community/Media'
@@ -22,9 +23,18 @@ export { CompanyDash } from 'pages/Dashboard/Company/CompanyDash'
 export { MarketProfessionals } from 'pages/Dashboard/Company/MarketProfessionals'
 export { Dashboard } from 'pages/Dashboard/Dashboard'
 export { DashboardLayout } from 'pages/Dashboard/DashboardLayout'
-export { DashProfessionalOffers } from 'pages/Dashboard/Professional/Offers'
+export { default as Offers } from 'pages/Dashboard/Professional/Offers'
 export { ProfessionalDash } from 'pages/Dashboard/Professional/ProfessionalDash'
 export { Error404 } from 'pages/Errors/Error404'
+export { AddInputDynamic } from 'pages/Form/AddInputDynamic'
+export { CVForm } from 'pages/Form/CVForm'
+export { ContactInfo } from 'pages/Form/ContactInfo'
+export { InfoForm } from 'pages/Form/InfoForm'
+export { ProfileImageForm } from 'pages/Form/ProfileImageForm'
+export { ProyectForm } from 'pages/Form/ProyectForm'
+export { UserForm } from 'pages/Form/UserForm'
+export { default as UserProfile } from 'pages/Form/UserProfile'
+export { UserProfileForm } from 'pages/Form/UserProfileForm'
 export { default as HomeEmpresa } from 'pages/HomeEmpresa'
 export { default as About } from 'pages/Info/About'
 export { default as Contact } from 'pages/Info/Contact'
@@ -32,10 +42,9 @@ export { default as InfoProfessionals } from 'pages/Info/InfoProfessionals'
 export { default as News } from 'pages/Info/News'
 export { OfferDetail } from 'pages/Info/OfferDetail'
 export { default as Offers } from 'pages/Info/Offers'
-export { default as Pricing } from 'pages/Info/pricing/Pricing'
 export { PricingEmpresa } from 'pages/Info/pricing/PricingEmpresa'
+export { default as PricingPage } from 'pages/Info/pricing/PricingPage'
 export { PricingProfesional } from 'pages/Info/pricing/PricingProfesional'
 export { Into } from 'pages/Into'
 export { Landing } from 'pages/Landing'
-export { default as UserForm } from 'pages/UserForm'
 export { default as Footer } from 'pages/footer/Footer'
